1.0.1

+++ RelatedChat/Classes/Chat/ChatView.m
@@ -28,7 +28,7 @@
 {
 	NSString *groupId;
 
-	int typingCounter;
+	NSInteger typingCounter;
 
 	FIRDatabaseReference *firebase1;
 	FIRDatabaseReference *firebase2;
